Allow seeding using global random package
Previously we required the user to create an instance of Random
and pass it around RandObj instances.
This did not map well on to the intended usage in many programs.

Instead, the user can now choose to not pass any Random instance to
a RandObj. In which case, it will use the Python random package.

This relies on the user controlling the seeding and random state of
the random package manually such that it is consistent across the
program and always accessed in a deterministic order.