Fix unusual issue around non-pure constraint
Constraints are evaluated early by `RandVar.create_randomizer()`.

This is unintuitive to a user. They might set up a constraint
which is non-pure, i.e. it relies on state outside of the constraint
problem. In which case, the constraint evaluating when it is added
to the problem might throw an exception.

Allow for this case by ignoring the exception when evaluating
constraints outside of calling `randomize()`.